refactor: move backlight backend out of backlight module
