+++ src/util/audio_backend.cpp
@@ -11,8 +11,6 @@
 
 namespace waybar::util {
 
-void noop() {}
-
 AudioBackend::AudioBackend(std::function<void()> on_updated_cb, private_constructor_tag tag)
     : mainloop_(nullptr),
       mainloop_api_(nullptr),
@@ -208,8 +206,7 @@ void AudioBackend::changeVolume(uint16_t volume, uint16_t min_volume, uint16_t m
   double volume_tick = static_cast<double>(PA_VOLUME_NORM) / 100;
   pa_cvolume pa_volume = pa_volume_;
 
-  volume = std::min(volume, max_volume);
-  volume = std::max(volume, min_volume);
+  volume = std::clamp(volume, min_volume, max_volume);
   pa_cvolume_set(&pa_volume, pa_volume_.channels, volume * volume_tick);
 
   pa_context_set_sink_volume_by_index(context_, sink_idx_, &pa_volume, volumeModifyCb, this);
